Retry connecting to the backend server if initial connection fails
The proxy now takes turns trying to connect to all backend sockets and sleeps
for 200ms if it is unable to do so. This allows the backend server to come up
up to 10 seconds after the client has authenticated.

It should be noted that using a connection timeout would not help with the
IPv6/IPv4 dual stack case.

Signed-off-by: Apollon Oikonomopoulos <apollon@noc.grnet.gr>