Handle socket errors, improve exception handling
Handle failure of socket operations (e.g., could not bind socket
due to address already in use), and inform the client accordingly.
To achieve this, the listening sockets get created by the main thread,
and passed to the greenlet that waits for client connections.

Also refactored the greenlet code to perform the RFB handshake
and the creation of forwarding greenlets as two distinct steps.
Any exceptions raised in greenlet context are caught and logged
appropriately.

Fix vapclient.py to exit with a non-zero status if the VNC
forwarding request fails.