Support dynamic port pool, encode reqs in JSON
Support dynamic allocation of client ports off a configurable port
pool, if the client so requests. The pool is initialized to the IANA
dynamic port range of 49152-65535 by default.

Use JSON encoding for requests and replies on the control channel.
If the client requests a dynamic port, by setting "source_port": 0,
the server returns the actual port used in the JSON reply.

Also fix vapclient.py:request_forwarding() to send requests and
receive replies in JSON, throwing Python exceptions in case of errors.