Fix possible race in returning source port to pool
First make sure all sockets for a connection are closed,
then return the source port number to the port pool.

This fixes a possible race where the greenlet gets preempted right after
returning a port to the pool, and the port is picked up from the pool by
another greenlet while there's still an open client socket using it.