Modify default port range for port pool
The default port range used for automatic port selection
was chosen to be the ephemeral port range. This was the wrong
thing to do, because it may conflict with ports used for outgoing
connections from the host on which vncauthproxy runs,
see /proc/sys/net/ipv4/ip_local_port_range on Linux.

We modify the default to be outside the ephemeral port range,
at 25000-30000.